+++ executor/seqtest/seq_executor_test.go
@@ -1082,63 +1082,3 @@ func (s *seqTestSuite) TestMaxDeltaSchemaCount(c *C) {
 	c.Assert(variable.GetMaxDeltaSchemaCount(), Equals, int64(2048))
 	tk.MustQuery("select @@global.tidb_max_delta_schema_count").Check(testkit.Rows("2048"))
 }
-
-func (s *seqTestSuite) TestRenameTable(c *C) {
-	c.Assert(failpoint.Enable("github.com/pingcap/tidb/meta/autoid/mockAutoIDChange", `return(true)`), IsNil)
-	defer func() {
-		c.Assert(failpoint.Disable("github.com/pingcap/tidb/meta/autoid/mockAutoIDChange"), IsNil)
-	}()
-	tk := testkit.NewTestKit(c, s.store)
-
-	tk.MustExec("create database rename1")
-	tk.MustExec("create database rename2")
-	tk.MustExec("create database rename3")
-	tk.MustExec("create table rename1.t (a int primary key auto_increment)")
-	tk.MustExec("insert rename1.t values ()")
-	tk.MustExec("rename table rename1.t to rename2.t")
-	// Make sure the drop old database doesn't affect the rename3.t's operations.
-	tk.MustExec("drop database rename1")
-	tk.MustExec("insert rename2.t values ()")
-	tk.MustExec("rename table rename2.t to rename3.t")
-	tk.MustExec("insert rename3.t values ()")
-	tk.MustQuery("select * from rename3.t").Check(testkit.Rows("1", "30001", "60001"))
-	// Make sure the drop old database doesn't affect the rename3.t's operations.
-	tk.MustExec("drop database rename2")
-	tk.MustExec("insert rename3.t values ()")
-	tk.MustQuery("select * from rename3.t").Check(testkit.Rows("1", "30001", "60001", "60002"))
-	tk.MustExec("drop database rename3")
-
-	tk.MustExec("create database rename1")
-	tk.MustExec("create database rename2")
-	tk.MustExec("create table rename1.t (a int primary key auto_increment)")
-	tk.MustExec("rename table rename1.t to rename2.t1")
-	tk.MustExec("insert rename2.t1 values ()")
-	result := tk.MustQuery("select * from rename2.t1")
-	result.Check(testkit.Rows("1"))
-	// Make sure the drop old database doesn't affect the t1's operations.
-	tk.MustExec("drop database rename1")
-	tk.MustExec("insert rename2.t1 values ()")
-	result = tk.MustQuery("select * from rename2.t1")
-	result.Check(testkit.Rows("1", "2"))
-	// Rename a table to another table in the same database.
-	tk.MustExec("rename table rename2.t1 to rename2.t2")
-	tk.MustExec("insert rename2.t2 values ()")
-	result = tk.MustQuery("select * from rename2.t2")
-	result.Check(testkit.Rows("1", "2", "30001"))
-	tk.MustExec("drop database rename2")
-
-	tk.MustExec("create database rename1")
-	tk.MustExec("create database rename2")
-	tk.MustExec("create table rename1.t (a int primary key auto_increment)")
-	tk.MustExec("insert rename1.t values ()")
-	tk.MustExec("rename table rename1.t to rename2.t1")
-	// Make sure the value is greater than autoid.step.
-	tk.MustExec("insert rename2.t1 values (100000)")
-	tk.MustExec("insert rename2.t1 values ()")
-	result = tk.MustQuery("select * from rename2.t1")
-	result.Check(testkit.Rows("1", "100000", "100001"))
-	_, err := tk.Exec("insert rename1.t values ()")
-	c.Assert(err, NotNil)
-	tk.MustExec("drop database rename1")
-	tk.MustExec("drop database rename2")
-}
